Refactor: consolidate edit messages.
