+++ cli/args.js
@@ -16,7 +16,7 @@ export function makeArgs(argv = process.argv) {
       engine: { type: "string", default: process.env["AILLY_ENGINE"] },
       model: { type: "string", default: process.env["AILLY_MODEL"] },
       plugin: { type: "string", default: process.env["AILLY_PLUGIN"] ?? "noop", },
-      context: { type: "string", default: process.env["AILLY_CONTEXT"] ?? "content", short: "c" },
+      context: { type: "string", default: process.env["AILLY_CONTEXT"], short: "c" },
       "template-view": { type: "string", default: process.env["AILLY_TEMPLATE_VIEW"] ? [process.env["AILLY_TEMPLATE_VIEW"]] : [], multiple: true },
       prompt: { type: "string", default: process.env["AILLY_PROMPT"], short: "p" },
       system: { type: "string", default: process.env["AILLY_SYSTEM"], short: "s" },
@@ -58,8 +58,8 @@ export function help() {
     --combined will force files to output as combined.
     -o, --out specify an output folder to work with responses. Defaults to --root. Will load responses from and write outputs to here, using .ailly file extensions.
     -c, --context conversation | folder | none
-      'conversation' (default) loads files from the root folder and includes them alphabetically, chatbot history style, before the current file when generating.
-      'folder' includes all files in the folder at the same level as the current file when generating.
+      'conversation' (default, unless --edit) loads files from the root folder and includes them alphabetically, chatbot history style, before the current file when generating.
+      'folder' includes all files in the folder at the same level as the current file when generating. Default with --edit.
       'none' includes no additional content (including no system context) when generating.
       (note: context is separate from isolated. isolated: true with either 'content' or 'folder' will result in the same behavior with either. With 'none', Ailly will send _only_ the prompt when generating.)
